rewrite iterating and rendering logic
Iterating thought the animation was based on the previous rendered
frame. In this new logic, the displayed frame is based on the elapsed
time (which takes into consideration the playback speed rate at each
display link trigger).

This way, completion detection it entirely based on the time spent
animating the image rather than relying on frame indices.

It also enables a new seeking method as well as negative playback
speed (reverse).